pyproject.toml
[project]
name = "west-standard-names"
version = "0.1.0"
description = "WEST tokamak standard names documentation using DD version 3.42.2"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.0.0",
    "mkdocs-macros-plugin>=1.0.0",
    "mike>=2.0.0",
    "pyyaml>=6.0",
    "pymdown-extensions>=10.0.0",
]
